Moving runner fixture into common conftest.py
